2026/06/22/17:38

+++ src/main/java/com/example/macaron/Store.java
@@ -2,20 +2,20 @@ package com.example.macaron;
 
 //パスワード、クーポン、自由記述が編集可能
 public class Store {
-    private final Long id;
+    private final Integer id;
     private final String name;
     private final String email;
     private String password;
     private final String address;
-    private final int phone_number;
+    private final Long phone_number;
     private final String payment;
-    private final int price_range;
+    private final Integer price_range;
     private final String genre;
     private String coupon;
     private String free_desc;
     
-    public Store(Long id, String name, String email, String password, String address, int phone_number, String payment,
-            int price_range, String genre, String coupon, String free_desc) {
+    public Store(Integer id, String name, String email, String password, String address, Long phone_number, String payment,
+            Integer price_range, String genre, String coupon, String free_desc) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -30,7 +30,7 @@ public class Store {
     }
    
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -50,7 +50,7 @@ public class Store {
         return address;
     }
 
-    public int getPhone_number() {
+    public Long getPhone_number() {
         return phone_number;
     }
 
@@ -58,7 +58,7 @@ public class Store {
         return payment;
     }
 
-    public int getPrice_range() {
+    public Integer getPrice_range() {
         return price_range;
     }
